+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 El formato sigue [Keep a Changelog](https://keepachangelog.com/es-ES/1.1.0/)
 y el proyecto usa [versionado semántico](https://semver.org/lang/es/).
 
+## [1.1.0] - 2026-08-29
+
+### Añadido
+
+- Localización automática en español e inglés para el plugin, las acciones, los
+  tooltips, los estados y los títulos mostrados en las teclas.
+
 ## [1.0.1] - 2026-08-29
 
 ### Cambiado